+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Polyverse Boost Automatic Application Modernization
 - Fix issue where some processing errors can be lost and misreported
 - Fix issue where files with '#' in the filename are not handled correctly
 - Fix issue where stale file cleanup is not run automatically on project analysis
+- Fix issue with disabled "Run Selected Analyses" button when launch via Command Palette
 
 ## Version 1.7.0: November 11th, 2023